hello-sam/tests/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
boto3==1.28.20
botocore==1.31.20; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0; python_version >= '3.7'
cryptography==41.0.3; python_version >= '3.7'
exceptiongroup==1.1.2; python_version < '3.11'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.7.0; python_version < '3.8'
iniconfig==2.0.0; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
moto==4.1.14
packaging==23.1; python_version >= '3.7'
pluggy==1.2.0; python_version >= '3.7'
pycparser==2.21
pytest==7.4.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
responses==0.23.3; python_version >= '3.7'
s3transfer==0.6.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version < '3.11'
types-pyyaml==6.0.12.11
typing-extensions==4.7.1; python_version < '3.8'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.3; python_version >= '3.7'
xmltodict==0.13.0; python_version >= '3.4'
zipp==3.15.0; python_version >= '3.7'
